+++ src/main/java/org/truffleruby/language/methods/ExceptionTranslatingNode.java
@@ -234,6 +234,18 @@ public class ExceptionTranslatingNode extends RubyNode {
             throw (AssertionError) throwable;
         }
 
+        if (throwable.getClass().getName().equals("com.oracle.truffle.api.vm.PolyglotIllegalStateException")) {
+            final String message;
+
+            if (throwable.getMessage().startsWith("Multi threaded access requested")) {
+                message = throwable.getMessage() + " Try running Ruby in single-threaded mode by using -Xsingle_threaded or --ruby.single_threaded.";
+            } else {
+                message = throwable.getMessage();
+            }
+
+            return coreExceptions().securityError(message, this);
+        }
+
         if (getContext().getOptions().EXCEPTIONS_PRINT_JAVA
                 || getContext().getOptions().EXCEPTIONS_PRINT_UNCAUGHT_JAVA) {
             throwable.printStackTrace();